Begin reverting of GValue code introduced with revision 21132 because of the huge memory usage to the code present when Xarchiver was first imported on svn.xfce.org. Fixed missed space in the rar command line string causing rar to fail when using a password. Fixed bug in the new dialog causing it not to detect file overwrite. Changed string in src/interface.c: SFX instead of Make SFX. Updated it.po Updated documentation. Issued make update-po.
(Old svn revision: 23324)

+++ src/archive.c
@@ -116,7 +116,9 @@ void xa_clean_archive_structure ( XArchive *archive)
 
 	if (archive->tmp != NULL)
 	{
-		unlink (archive->tmp);
+		if ( strstr (archive->tmp , "/tmp" ) )
+			unlink (archive->tmp);
+
 		g_free(archive->tmp);
 		archive->tmp = NULL;
 	}
